intial commit, fixed
