+++ stock-trainer-front-end/src/components/NavBreadcrumbs.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import { Breadcrumb } from 'semantic-ui-react'
+
+const NavBreadcrumbs = () => (
+  <Breadcrumb  size='large'>
+    <Breadcrumb.Section link>Home</Breadcrumb.Section>
+    
+    
+  </Breadcrumb>
+)
+
+export default NavBreadcrumbs
