+++ stock-trainer-front-end/src/components/TopBar.jsx
@@ -1,16 +1,25 @@
-import React, { Component } from 'react'
-import Logo from './Logo';
-import SignedOutContainer from './SignedOutContainer';
-import SignInContainer from './SignInContainer';
+import React, { Component } from "react";
+import Logo from "./Logo";
+import SignOutContainer from "./SignInContainer";
+import SignInContainer from "./SignOutContainer";
+import { Grid } from "semantic-ui-react";
+import NavBreadcrumbs from "./NavBreadcrumbs";
 
 export default class TopBar extends Component {
   render() {
     return (
-      <div className="topBarContainer">
-        <Logo/>
-        <SignInContainer />
-        {/* <SignedOutContainer currentUser = {this.props.currentUser}/> */}
-      </div>
-    )
+      <Grid padded>
+        <Grid.Column width={5}>
+          <Logo />
+        </Grid.Column>
+        <Grid.Column floated="right" width={4}>
+          {this.props.signedInState ? (
+            <SignOutContainer currentUser={this.props.currentUser} />
+          ) : (
+            <SignInContainer />
+          )}
+        </Grid.Column>
+      </Grid>
+    );
   }
 }
